+++ src/System.IO/tests/Stream/Stream.cs
