+++ src/System.IO/tests/StreamReader/StreamReader.cs
